+++ docs/changelog.md
@@ -2,6 +2,7 @@
 - Refactor to separate omnibar and search
 - Add CTRL+P for omnibar focus
 - Highlight currently selected item
+- Make Escape on omnibar return focus to last focused item
 
 # Version 3.5.0: More useful search 
 - Fix minor focus bug